// modules/configurator/models/FinalConfigModel.ts
export interface FinalConfig {
    id: string;
    name: string;
    year: number;
    color: string;
    colorFull: string;
    wheels: string;
    wheelsFull: string;
    interior: string;
    interiorFull: string;
    carType: string;
    price: number;
    colorPrice: number;
    wheelsPrice: number;
    interiorPrice: number;
  }